third_party/WebKit/LayoutTests/fast/dom/shadow/exposed-object-within-shadow-expected.txt

CONSOLE WARNING: line 19: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Shadow DOM inside an <object> and a <div>
PASS child.name in document is false
PASS child.name in window is false
PASS obj.name in document is true
Shadow DOM attached to <div> with an <object> sibling (next).
PASS child.name in document is false
PASS child.name in window is false
PASS obj.name in document is true
Shadow DOM attached to <div> with an <object> sibling (previous).
PASS child.name in document is false
PASS child.name in window is false
PASS obj.name in document is true
Shadow DOM attached to <div> inside a <param> inside an <object>.
PASS obj.name in document is true
PASS child.name in document is false
PASS child.name in window is false
PASS obj.name in document is true
PASS successfullyParsed is true

TEST COMPLETE
